+++ examples/decoding.rs
@@ -104,8 +104,8 @@ fn process_frame(yuv_frame: MediaFrame<u8>) -> Result<()> {
     let rgb_frame = yuv_frame.convert_yuv420p_to_rgb24()?;
 
     let img: ImageBuffer<Rgb<u8>, Vec<u8>> = ImageBuffer::from_raw(
-        yuv_frame.width as u32,
-        yuv_frame.height as u32,
+        yuv_frame.width,
+        yuv_frame.height,
         rgb_frame
             .data
             .as_packed()
